requirements.txt
lxml>=4.9.3
pyodbc>=4.0.39
xlsxwriter>=3.1.9
openpyxl>=3.1.2
PyMySQL>=1.1.0
mysql-connector-python>=8.2.0